pyproject.toml
[tool.poetry]
name = "flask-006"
version = "0.1.0"
description = ""
authors = ["Mapc <mapctrigotixd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
psycopg2-binary = "^2.9.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
